# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DCP Platform API
  version: 5.0.0
  description: |
    DCP is Saudi Arabia's first decentralized GPU compute marketplace.
    Providers register NVIDIA GPUs, install the Python daemon, and earn SAR.
    Renters submit compute jobs (LLM inference, image generation, training) that
    run on provider hardware. DCP takes 25% fee; providers earn 75%.

    ## Currency
    All monetary amounts are in **halala** (1 SAR = 100 halala) unless the field
    name ends in `_sar`.

    ## Auth
    Three auth schemes are used depending on the caller role:

    | Role | Scheme | How |
    |------|--------|-----|
    | Provider | `ProviderApiKey` | `?key=` query param or `x-provider-key` header |
    | Renter | `RenterApiKey` | `?key=` query param or `x-renter-key` header |
    | Admin | `AdminToken` | `x-admin-token` header |

    ## Rate Limits
    | Endpoint | Limit |
    |----------|-------|
    | `POST /providers/register` | 5 / IP / hour |
    | `POST /providers/heartbeat` | 4 / IP / minute |
    | `POST /renters/register` | 5 / IP / hour |
    | `POST /jobs/submit` | 10 / renter API key / minute |
    | `POST /payments/topup` | 10 / IP / minute |
    | `GET /api/providers/marketplace` | 30 / IP / minute |
    | `GET /api/admin/*` | 100 / IP / minute |
    | All other `/api/*` | 300 / IP / minute |

  contact:
    name: DCP DevRel
    url: https://dcp.sa/docs
  x-env-vars:
    DC1_ADMIN_TOKEN:
...
      required: true
    DC1_HMAC_SECRET:
...
      required: true
...
servers:
...
components:
  securitySchemes:
    ProviderApiKey:
      type: apiKey
      in: query
      name: key
      description: Provider API key. Prefix `dcp-provider-`. Also accepted as `x-provider-key` header.
    RenterApiKey:
      type: apiKey
      in: query
      name: key
      description: Renter API key. Prefix `dcp-renter-`. Also accepted as `x-renter-key` header.
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: Admin token set via DC1_ADMIN_TOKEN env var. Also accepted as Bearer token.
...
paths:
...
  /api/providers/register:
    post:
      tags: [Providers]
      summary: Register a new GPU provider
      description: |
        Creates a new provider account and returns an API key and installer URL.
        Rate limited to 5 requests per IP per hour.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, gpu_model, os]
              properties:
                name:
                  type: string
                  example: Ahmed Al-Rashidi
                email:
                  type: string
                  format: email
                  example: ahmed@example.sa
                gpu_model:
                  type: string
                  example: NVIDIA RTX 3090
                os:
                  type: string
                  enum: [Windows, Linux, Mac]
                  example: Linux
                phone:
                  type: string
                  nullable: true
                resource_spec:
                  type: object
                  nullable: true
                  description: Ocean-style resource specification (cpu_count, ram_gb, etc.)
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  provider_id:
                    type: integer
                  api_key:
                    type: string
                    example: dcp-provider-a1b2c3d4e5f6...
                  installer_url:
                    type: string
                  message:
                    type: string
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: A provider with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/providers/login-email:
    post:
      tags: [Providers]
      summary: Retrieve API key by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: API key returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  api_key:
                    type: string
                  provider:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      email:
                        type: string
                      gpu_model:
                        type: string
                      status:
                        type: string
        '404':
          description: No provider account found with this email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/providers/heartbeat:
    post:
      tags: [Providers]
      summary: Provider daemon heartbeat
      description: |
        Sent every 30 seconds by the daemon. Updates GPU telemetry, sets provider
        status to `online`, recomputes reputation score, and returns job assignment
        signals and update availability.

        Rate limited to 4 per IP per minute.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [api_key]
              properties:
                api_key:
                  type: string
                  example: dcp-provider-a1b2c3d4e5f6...
                gpu_status:
                  type: object
                  description: GPU telemetry snapshot from nvidia-smi
                  properties:
                    gpu_name:
                      type: string
                    gpu_vram_mib:
                      type: integer
                    driver_version:
                      type: string
                    gpu_util_pct:
                      type: number
                    temp_c:
                      type: number
                    power_w:
                      type: number
                    free_vram_mib:
                      type: integer
                    daemon_version:
                      type: string
                    python_version:
                      type: string
                    os_info:
                      type: string
                    compute_capability:
                      type: string
                    cuda_version:
                      type: string
                    gpu_count:
                      type: integer
                    all_gpus:
                      type: array
                      items:
                        type: object
                uptime:
                  type: number
                  description: Daemon uptime in seconds
                provider_ip:
                  type: string
                provider_hostname:
                  type: string
                cached_models:
                  type: array
                  items:
                    type: string
                resource_spec:
                  type: object
      responses:
        '200':
          description: Heartbeat acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  update_available:
                    type: boolean
                    description: True if daemon version is below minimum required
                  min_version:
                    type: string
                    example: "3.3.0"
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/renters/register:
    post:
      tags: [Renters]
      summary: Register a new renter account
      description: |
        Creates renter account and credits 1000 halala (10 SAR) welcome balance.
        Rate limited to 5 per IP per hour.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                  example: Fatima Al-Saud
                email:
                  type: string
                  format: email
                  example: fatima@example.sa
                organization:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  renter_id:
                    type: integer
                  api_key:
                    type: string
                    example: dcp-renter-a1b2c3d4e5f6...
                  message:
                    type: string
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/renters/topup:
    post:
      tags: [Renters]
      summary: Direct balance top-up (dev/legacy — use /api/payments/topup for production)
      description: |
        Directly credits halala to renter balance. Accepts either `amount_halala`
        or `amount_sar`. Max 1000 SAR per transaction. Rate limited to 10/min.
      security:
        - RenterApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount_halala:
                  type: integer
                  description: Amount in halala (mutually exclusive with amount_sar)
                amount_sar:
                  type: number
                  description: Amount in SAR (mutually exclusive with amount_halala)
      responses:
        '200':
          description: Balance credited
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  topped_up_halala:
                    type: integer
                  topped_up_sar:
                    type: number
                  new_balance_halala:
                    type: integer
                  new_balance_sar:
                    type: number
  /api/renters/login-email:
    post:
      tags: [Renters]
      summary: Retrieve API key by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: API key returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  api_key:
                    type: string
                  renter:
                    $ref: '#/components/schemas/Renter'
...
  /api/jobs/submit:
    post:
      tags: [Jobs]
      summary: Submit a compute job
      description: |
        Validates renter balance, deducts estimated cost as a pre-pay hold, assigns
        provider, and queues the job.

        Enforced controls:
        - Rate limit: **10 submissions/min per renter API key**
        - Balance gate: renter `balance_halala` must be `> 0`
        - Quotas: `daily_jobs_limit` and `monthly_spend_limit_halala` (from `renter_quota`)

        **Supported job types and cost rates (halala/min):**
        | Type | Rate |
        |------|------|
        | `llm-inference` | 15 |
        | `training` | 25 |
        | `rendering` | 20 |
        | `image_generation` | 20 |
        | `vllm_serve` | 20 |
        | `custom` | 10 |
      security:
        - RenterApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [job_type]
              properties:
                job_type:
                  type: string
                  enum: [llm-inference, llm_inference, training, rendering, image_generation, vllm_serve, custom]
                  example: image_generation
                duration_minutes:
                  type: integer
                  default: 10
                  description: Estimated runtime. Used for pre-pay cost calculation.
                task_spec:
                  type: object
                  description: |
                    Job-specific parameters. For supported types, the API auto-generates
                    the task script from these params (no custom Python needed).

                    **image_generation params:**
                    - `prompt` (string)
                    - `negative_prompt` (string)
                    - `steps` (int, 5–100)
                    - `width` / `height` (int, 256–1024)
                    - `seed` (int)
                    - `model` (enum — see allowed SD models)

                    **llm-inference params:**
                    - `prompt` (string)
                    - `max_tokens` (int, 32–4096)
                    - `temperature` (float, 0.1–2.0)
                    - `model` (enum — see allowed LLM models)
                gpu_requirements:
                  type: object
                  nullable: true
                  description: Optional GPU filter (min_vram_gb, gpu_model, etc.)
                max_duration_seconds:
                  type: integer
                  default: 1800
                  description: Hard timeout. Job is killed if it exceeds this.
                priority:
                  type: integer
                  enum: [1, 2, 3]
                  default: 2
                  description: "1=high, 2=normal, 3=low"
      responses:
        '201':
          description: Job submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  job_id:
                    type: string
                  status:
                    type: string
                    enum: [pending, queued]
                  provider_id:
                    type: integer
                    nullable: true
                  cost_halala:
                    type: integer
                  new_balance_halala:
                    type: integer
        '400':
          description: Invalid request (missing job_type, etc.)
        '402':
          description: Insufficient balance or zero wallet balance
        '429':
          description: Renter submission rate limit or quota exceeded
        '503':
          description: No providers available
...
  /api/payments/topup:
    post:
      tags: [Payments]
      summary: Initiate a Moyasar SAR payment
      description: |
        Creates a Moyasar payment and returns a hosted checkout URL. Renter must
        be redirected to `checkout_url` to complete payment. The webhook confirms
        and credits the balance. Rate limited to 10/min.
      security:
        - RenterApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_sar]
              properties:
                amount_sar:
                  type: number
                  minimum: 1
                  maximum: 10000
                  example: 100
                source_type:
                  type: string
                  enum: [creditcard, mada, applepay]
                  default: creditcard
                callback_url:
                  type: string
                  format: uri
                  nullable: true
                  description: Override default redirect URL after payment
      responses:
        '200':
          description: Payment initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  payment_id:
                    type: string
                  amount_sar:
                    type: number
                  amount_halala:
                    type: integer
                  status:
                    type: string
                  checkout_url:
                    type: string
                    format: uri
                  message:
                    type: string
        '503':
          description: Payment gateway not configured
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  sandbox_hint:
                    type: string
  /api/payments/topup-sandbox:
    post:
      tags: [Payments]
      summary: Dev-only sandbox balance top-up
      description: |
        Directly credits balance without Moyasar. Only available when
        `MOYASAR_SECRET_KEY` is **not** set (dev/staging environments). Disabled in production.
      security:
        - RenterApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_sar]
              properties:
                amount_sar:
                  type: number
                  minimum: 0.01
                  maximum: 10000
      responses:
        '200':
          description: Balance credited (sandbox)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  sandbox:
                    type: boolean
                  payment_id:
                    type: string
                  amount_sar:
                    type: number
                  credited_halala:
                    type: integer
                  new_balance_sar:
                    type: number
                  new_balance_halala:
                    type: integer
        '403':
          description: Sandbox disabled in production
  /api/payments/webhook:
    post:
      tags: [Payments]
      summary: Moyasar payment webhook
      description: |
        Receives payment events from Moyasar. Verifies HMAC-SHA256 signature
        (`x-moyasar-signature` header). Credits balance on `paid` event.
        Must return 2xx or Moyasar will retry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: Moyasar payment ID
                status:
                  type: string
                  enum: [initiated, paid, failed, refunded]
                amount:
                  type: integer
                  description: Amount in halala
                amount_refunded:
                  type: integer
                  nullable: true
      responses:
        '200':
          description: Webhook acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                  action:
                    type: string
                    enum: [balance_credited, marked_failed, refund_processed, already_processed, ignored_unknown, status_updated]
        '401':
          description: Invalid HMAC signature
...
  /api/admin/providers:
    get:
      tags: [Admin]
      summary: List all providers
      security:
        - AdminToken: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          description: Page number (0 = all, 1+ = paginated)
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            maximum: 200
        - in: query
          name: search
          schema:
            type: string
          description: Filter by name, email, or GPU model
        - in: query
          name: status
          schema:
            type: string
            enum: [online, offline, registered, suspended]
      responses:
        '200':
          description: Providers list
    post:
      tags: [Admin]
      summary: Bulk provider operations
      security:
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Bulk operation result
...
  /api/admin/payments:
    get:
      tags: [Admin]
      summary: List all platform payments
      security:
        - AdminToken: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Payments list
...